Populates External Media Folder Dropdown via stubbed API request
Previously folder data to populate the dropdown was passed down statically into the component from the MediaLibraryExternalHeader. Updates to create a Flux Store implementation `MediaFolderStore` and a controller view `MediaFolderData` to manage requesting the folder data from the WPCom API and storing in state. This mirrors the established Flux patterns within the Media Library. The WPCom API request is currently a stub because it is yet to be implemented.