Adds Folder view to External Media list
Previous Media view was only capable of handling photo media items. Adds Folder Item and updates List Item to handle double click to “enter” a folder. Media Query “path” is set as the folder ID value in order that WPCom integration requests either that specific Album (returning the photos within it) or the “root” (returning all the Albums in the user’s library).