Fixs Folder thumbnails to occupy full box height
Previously the thumbnail images used as the cover of each Album would not occupy the full space of the item. This was particular obvious when a landscape ratio image was used as the Album thumbnail. Fix this by using background-image to force the image to occupy the full item space, but retain the original <img> tag in the source for accessiblity reasons.